services/web_service/requirements.txt
Flask
Flask-WTF
httpx
python-dotenv